Improve failure diagnostics
When CUDA compilation fails, the requested PTX file won't exist, but the
only means of discovering that is to see that your generated C file
contains an error message from bin2c.  Explicitly test for the existence
of the expected PTX file and fail the CMake script with a message and
write a #error directive in the generated C file.