Ensure that page mapping vector is always sorted
When we split the ID space into "backed" and "unbacked" partitions, we
introduced two methods for allocating page ids and the returned ids
aren't necessarily contiguous anymore.  When looking up a request
handler, this could cause us to miss the page mapping entry
corresponding to a page id because the vector of mappings wasn't sorted
anymore.  Ensure that we always insert the new mapping into the correct
position in the vector.