Drop Resource structure in proxy instances
Trying to record the appropriate index for an page in the Resource
structure fails as soon as we modify the set of proxies by removing one.
Once a proxy has been removed, all the stored indices in the Resource
structures are wrong.

So drop the Resource structure altogether and keep the list of proxy
page ids sorted by page id and use the offset from the beginning of
the proxy page id list as the index at which to drop the other array
entries.

Rework test_proxyInstances to use matchers to validate the inputs of
function calls optixAccelComputeMemoryUsage and optixAccelBuild.  Add a
test case that covers the modification of the proxy set in a non-trivial
way to catch the indexing errors.