Image tests have a resource target that updates gold files and resources
Since image tests are run against gold images in the binary directory,
the test can become stale when the gold image is updated in the source
directory.  Add a custom target that sets up dependencies between the
file in the binary directory and the source directory and updates the
file accordingly.  This was previously only done once at configure time.

Add a FOLDER parameter to add_image_test that allows you to specify the
IDE folder of the custom target that keeps the files up to date.

Add a RESOURCES parameter to add_image_test that allows you to specify
additional test resource files that need to be updated in the build
directory under ${CMAKE_BINARY_DIR}/tests/resources/${target}.