Add ProgramGroupDescBuilder member overloads and ISAHCH member
Optix program group descriptions reference modules and entry point
function names as pairs in the OptixProgramGroupSingleModule structure.
Unfortunately the hit group record is composed of 6 discrete members
instead of 3 instances of this struct, while the raygen and miss members
are a single instance of this struct.  The module and entry point
function name always go together, with the entry function name possibly
bieng nullptr for builtin intersection modules.  Add builder overloads
to take the struct instead of the {module,function name} pair of
arguments.  There's no need to introduce a C++ wrapper for
OptixProgramGroupSingleModule as the rules for aggregate initialization
cover the desired semantics properly.

Add a hitGroupISAHCH member that lets you specify all three module/function
pairs in a hitgroup record.  Always set the hitgroup record members in
the order of IS, AH and then CH.

Add tests for the ProgramGroupDescBuilder to cover all the variants and
members.

Rename the existing hitgroup builder members to follow the IS, AH, CH
entry point order.