Ensure consistent OptixModule printing in tests
Inserting a void* value into a stream is not defined by the Standard
Library.  Implementations are probably relying on the %p specifier,
which is implementation defined.  The end result is that some platforms
would print uppercase hex digits and some platforms would print
lowercase hex digits, making the tests fragile.  OptixModule is an
opaque pointer, so to insert an OptixModule value into a stream, we
convert it to an unsigned long long via memcpy (aka C++20 bit_cast<>)
and print the value as a hex string in lowercase with leading zeros.