Add vcpkg dependency support for stb
Use the suggested vcpkg find_package recipe and don't do FetchContent.

Since Stb is a header-only library with a mechanism for compiling the
implementation once, use that mechanism to build a static library
StbImage that contains the image reading and writing implementations.
Do this whether we are using vcpkg or FetchContent to obtain Stb.

With FetchContent, get the whole repo at the same hash as the recipe in
vcpkg for consistency, instead of fetching two files at an unknown
revision from the master branch.