Instead of including the vcpkg toolchain, tell CMake to do it
The first call to project() will include a toolchain file, if any is
specified.  So instead of trying to include it ourselves (we're being
called before project), set CMAKE_TOOLCHAIN_FILE if not already set and
let CMake include it at the proper time.

Also guard this script against being called multiple times.