Decouple pbrt API from pbrt scene loader
The scene loader only exposes parseFile and parseString, whereas the API
implementation covers every pbrt API entry point to build the scene
description.  The abstract API shouldn't be coupled to the scene
description exposed by the scene loader.  The DemandPbrtScene sample is
now coupled to otk::pbrt::SceneLoader and not the abstract API.