An ObjectInstance has a transform, but not an ObjectDefinition
An Object is a bag of transformed shapes; the bounds of the Object
is the union of the bounds of the transformed shapes.  An Object doesn't
have a transform, but an ObjectInstance does.  Therefore when building
ObjectDefinitions from the scene description, the current transform
doesn't apply to the ObjectDefinition.  The current transform has
already been applied to the contained shapes when they are defined.